+++ controller/patterns/sweep_pattern.py
@@ -6,8 +6,8 @@ import random
 class SweepPattern(Pattern):
     def __init__(self):
         super().__init__()
-        self.params.color = np.array([255, 255, 255])
-        self.params.background_color = np.array([0, 0, 0])
+        self.params.color = np.array([255, 255, 255], dtype=np.uint8)
+        self.params.background_color = np.array([0, 0, 0], dtype=np.uint8)
         self.params.decay_random = True
         self.params.decay_param = 0.5
         self.params.sweep_speed = 0.3
